refactor: UserService 인터페이스로 분리
