+++ app/src/main/java/com/techcourse/service/AppUserService.java
@@ -1,0 +1,45 @@
+package com.techcourse.service;
+
+import com.interface21.jdbc.core.TransactionTemplate;
+import com.techcourse.dao.UserDao;
+import com.techcourse.dao.UserHistoryDao;
+import com.techcourse.domain.User;
+import com.techcourse.domain.UserHistory;
+import java.util.NoSuchElementException;
+import javax.sql.DataSource;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class AppUserService implements UserService {
+
+    private static final Logger log = LoggerFactory.getLogger(AppUserService.class);
+
+    private final UserDao userDao;
+    private final UserHistoryDao userHistoryDao;
+
+    public AppUserService(final UserDao userDao, final UserHistoryDao userHistoryDao) {
+        this.userDao = userDao;
+        this.userHistoryDao = userHistoryDao;
+    }
+
+    @Override
+    public User findById(final long id) {
+        return userDao.findById(id)
+            .orElseThrow(() -> new NoSuchElementException("id에 해당하는 유저를 찾을 수 없습니다."));
+    }
+
+    @Override
+    public void save(final User user) {
+        userDao.insert(user);
+    }
+
+    @Override
+    public void changePassword(final long id, final String newPassword, final String createBy) {
+        final var user = findById(id);
+        log.debug("user password change: {}", user);
+        user.changePassword(newPassword);
+
+        userDao.update(user);
+        userHistoryDao.log(new UserHistory(user, createBy));
+    }
+}
