+++ app/src/main/java/com/techcourse/service/UserService.java
@@ -1,48 +1,10 @@
 package com.techcourse.service;
 
-import com.interface21.jdbc.core.TransactionTemplate;
-import com.techcourse.dao.UserDao;
-import com.techcourse.dao.UserHistoryDao;
 import com.techcourse.domain.User;
-import com.techcourse.domain.UserHistory;
-import java.sql.Connection;
-import java.util.NoSuchElementException;
-import javax.sql.DataSource;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-public class UserService {
+public interface UserService {
 
-    private static final Logger log = LoggerFactory.getLogger(UserService.class);
-
-    private final TransactionTemplate transactionTemplate;
-    private final UserDao userDao;
-    private final UserHistoryDao userHistoryDao;
-
-    public UserService(final DataSource dataSource, final UserDao userDao, final UserHistoryDao userHistoryDao) {
-        this.userDao = userDao;
-        this.userHistoryDao = userHistoryDao;
-        this.transactionTemplate = new TransactionTemplate(dataSource);
-    }
-
-    public User findById(final long id) {
-        return userDao.findById(id)
-            .orElseThrow(() -> new NoSuchElementException("id에 해당하는 유저를 찾을 수 없습니다."));
-    }
-
-    public void insert(final User user) {
-        userDao.insert(user);
-    }
-
-    public void changePassword(final long id, final String newPassword, final String createBy) {
-        transactionTemplate.execute(() -> {
-            final var user = findById(id);
-            log.debug("user password change: {}", user);
-            user.changePassword(newPassword);
-
-            userDao.update(user);
-            userHistoryDao.log(new UserHistory(user, createBy));
-            return null;
-        });
-    }
+    User findById(final long id);
+    void save(final User user);
+    void changePassword(final long id, final String newPassword, final String createdBy);
 }
